chore: restructuring
